Added tests and functions

+++ scripts/filtering_data.py
@@ -44,7 +44,7 @@ class filter_data:
         """
         Filter a minimum number of tickets sold.
         """
-        return self.df[self.df['Tickets Sold'] > int(tickets_sold)]
+        return self.df[self.df['Tickets Sold'] > float(tickets_sold)]
     
         #ARREGLAR
